고객 적립금에서 오버 플로우 발생되는 이슈 해결

+++ app/src/main/java/com/moaplanet/gosingadmin/main/qrpayment/model/QrCodeViewModel.java
@@ -85,8 +85,8 @@ public class QrCodeViewModel extends BaseViewModel {
         if (inputSavePrice.getValue() != null && reserveRatio.getValue() != null) {
             String tempSavePrice = inputSavePrice.getValue().replace(",", "");
             if (!tempSavePrice.equals("")) {
-                int savePriceResilt = Integer.valueOf(tempSavePrice) * reserveRatio.getValue() / 100;
-                saveMoney.setValue(StringUtil.convertCommaPrice(savePriceResilt));
+                long savePriceResilt = Long.valueOf(tempSavePrice) * reserveRatio.getValue() / 100;
+                saveMoney.setValue(StringUtil.convertCommaPrice((int)savePriceResilt));
             }
         }
     }
